FC offset cal rev

+++ lib_timesync.py
@@ -70,7 +70,8 @@ def on_message(client, userdata, msg):
         now = float( dt.timestamp( dt.now() ) )        
         if rx_msg.time_unix_usec is not 0:
             monitor.fc_time = float( rx_msg.time_unix_usec / 1e6 )
-            monitor.fc_offset = int( ( (monitor.fc_time + monitor.fc_lt) - now ) * 1000 )
+            # monitor.fc_offset = int( ( (monitor.fc_time + monitor.fc_lt) - now ) * 1000 )
+            monitor.fc_offset = ((monitor.fc_time + monitor.fc_lt) - now ) * 1000
             print('Time calculation')
         else:
             monitor.fc_time = 0
